Fix error in missionsSlice.js file
